refactor: update route imports and fix build
